# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_getUsers_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getMe_v1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user information",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/change-password": {
      "patch": {
        "operationId": "UsersController_changePassword_v1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid password"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/profile": {
      "patch": {
        "operationId": "UsersController_updateUserProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/students/academic-info": {
      "get": {
        "operationId": "StudentProfilesController_getStudentsAcademicInfo_v1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentAcademicInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "Students academic info not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Roles: organization] Get students academic information",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/api/v1/students": {
      "get": {
        "operationId": "StudentProfilesController_uploadStudentProfiles_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Students"
        ]
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "operationId": "StudentProfilesController_getStudentDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Roles: organization] Get student detail by ID",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/api/v1/psychologists/{id}/availability": {
      "get": {
        "operationId": "PsychologistsController_getPsychologistAvailability_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Psychologists"
        ]
      }
    },
    "/api/v1/psychologists/{id}/details": {
      "get": {
        "description": "Get comprehensive psychologist details including availability, specialties, and booking info",
        "operationId": "PsychologistsController_getPsychologistDetails_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed psychologist information"
          }
        },
        "summary": "Get detailed psychologist information",
        "tags": [
          "Psychologists"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ruang Diri API",
    "description": "API documentation for Ruang Diri application, all the services belongs to PT. Wong Makmur Sejati",
    "version": "1.0",
    "contact": {
      "name": "Support Team",
      "url": "https://ruangdiri.id",
      "email": "support@ruangdiri.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://ruangdiri.id/terms"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token here"
      }
    },
    "schemas": {
...
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Jl. Sudirman No. 123, Jakarta"
          },
          "phone": {
            "type": "string",
            "example": "+62 812-3456-7890"
          },
          "profilePicture": {
            "type": "string",
            "example": "https://example.com/profile.jpg"
          },
          "fullName": {
            "type": "string",
            "example": "SMA Negeri 1 Jakarta"
          }
        },
        "required": [
          "address",
          "phone",
          "profilePicture",
          "fullName"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the user",
            "example": "9b3ac4fd-7790-40d1-9e8d-174a250b6799"
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "school@ruangdiri.com"
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the user",
            "example": "SMAN 123 3211"
          },
          "role": {
            "description": "User role",
            "example": [
              "organization",
              "student",
              "employee",
              "psychologist",
              "client"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organization": {
            "description": "Organization details if user is an organization",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrganizationDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "The current password of the user",
            "example": "currentPassword123"
          },
          "newPassword": {
            "type": "string",
            "description": "The new password to set",
            "example": "newSecurePassword456"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of the password change operation",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Message about the password change operation",
            "example": "Password changed successfully"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "StudentAcademicInfoDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Information about grades and classrooms of students",
            "example": {
              "classrooms": [
                "X",
                "XI",
                "XII"
              ],
              "grades": [
                "A",
                "B",
                "C",
                "D"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
      "StudentDetailDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "Detailed information about a student",
            "example": {
              "id": "63d4405c-d2fd-4bd8-8afe-90bb67cbbf5c",
              "email": "student13@example.com",
              "profilePicture": "https://improved-mammal-humane.ngrok-free.app/uploads/students/profilePicture-1747900509537-783722953.jpeg",
              "fullName": "Ursasi",
              "role": "student",
              "studentProfile": {
                "grade": "D",
                "classroom": "XI",
                "gender": "female",
                "nis": "11335789",
                "screeningStatus": "stable",
                "counselingStatus": false,
                "iqScore": 190,
                "guardianName": "Parent 13",
                "guardianContact": null,
                "birthDate": "2009-01-01T00:00:00.000Z",
                "birthPlace": "City Center",
                "iqCategory": "very_below_average"
              },
              "mentalHealthHistories": [
                {
                  "id": "951df749-c5b7-4f13-9bb0-0d3a4005ef38",
                  "type": "screening",
                  "date": "2025-03-22T07:37:56.211Z",
                  "status": "stable",
                  "notes": null
                },
                {
                  "id": "3a13391a-bc68-4797-8c91-83ce5d7f4e40",
                  "type": "screening",
                  "date": "2025-04-15T07:30:59.466Z",
                  "status": "not_screened",
                  "notes": null
                },
                {
                  "id": "d19eb847-345f-41be-8112-52c13f8f908a",
                  "type": "screening",
                  "date": "2025-03-26T07:30:59.466Z",
                  "status": "stable",
                  "notes": null
                }
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      }
    }
  }
}